+++ CHANGES.md
@@ -1,5 +1,16 @@
 #### [unreleased]
 
+#### 12.2.0 / 2023-04-20
+* update `freemius/wordpress-sdk`
+* update `afragen/wp-dismiss-notice`
+* don't save unused data from `API_Common::parse_release_asset()`
+* don't use Freemius uninstall, use previous `uninstall.php`
+* more PHP 8.2 compatibility
+* composer update
+* update `REST_API::get_plugins_api_data()` to return response without download link using boolean value in `download` query arg
+* hide Freemius menus with `gu_hide_settings` filter
+* more specific hiding of Git Updater settings
+
 #### 12.1.3 / 2023-03-20
 * improved setting/default of `$options['bypass_background_processing']`
 * improved setting/default of `$options['branch_switch']`